requirements.txt
beautifulsoup4==4.12.2
python-docx==0.8.11
PyQt5==5.15.9
lxml==4.9.3 